# marketing-ai/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: FoodTab Marketing AI — API v1
  version: "1.0.0"
  summary: Rozhraní pro n8n, FoodTab Řízení, cron a příchozí webhooky.
  description: >
    Smlouva rozhraní v1. Vlastní obrazovky aplikace tohle API nepoužívají —
    čtou data v server komponentách a zapisují server actions. API v1 je tu
    pro cizí systémy a pro automatizaci. Část cest je zatím jen smlouva a
    v aplikaci je obslouží obrazovka; u každé takové je to napsané
    v popisu. Chyby vracejí objekt Chyba s českou hláškou a nikdy
    neobsahují tajemství. Peníze jsou celé haléře jako integer, časy jsou
    v ISO 8601 s pásmem, provozní data jako date.
  contact:
    name: FoodTab
    url: https://foodtab.cz
servers:
  - url: https://marketing.foodtab.cz
    description: Ostrý provoz
  - url: http://localhost:3000
    description: Vývoj
...
paths:
...
  /api/v1/obsah/{id}/navrh:
    post:
      tags:
        - Obsah
      operationId: navrhnoutObsah
      summary: Nechat AI navrhnout varianty textu a storyboard
      description: Implementováno v aplikaci (přihlášení, X-Cron-Secret nebo podpis HMAC z n8n). Volá ji workflow content-generation.json v n8n, proto přijímá i podpis HMAC. Model dostane jen brief, brand kit a potvrzené menu — nikdy mzdy, docházku, kontakty ani zálohy. Návrh se uloží jako nová verze, publikovat se nedá bez schválení člověkem. Když je připojený jen demo návrhář, je výsledek viditelně označený jako mock.
      security:
        - cookieSession: []
        - webhookHmac: []
      parameters:
        - $ref: "#/components/parameters/IdVCeste"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                contentItemId:
                  type: string
                  format: uuid
                  description: Nepovinné, musí souhlasit s ID v cestě.
                duvod:
                  type: string
      responses:
        "200":
          description: Návrh vytvořen.
          content:
            application/json:
              schema:
                type: object
                properties:
                  versionId:
                    type: string
                    format: uuid
                  isMock:
                    type: boolean
                  navrh:
                    $ref: "#/components/schemas/AiNavrh"
        "401":
          $ref: "#/components/responses/Nepovoleno"
        "403":
          $ref: "#/components/responses/Zakazano"
        "404":
          $ref: "#/components/responses/Nenalezeno"
...
  /api/v1/integrace/test-vse:
    post:
      tags:
        - Integrace
      operationId: otestovatVsechnaPripojeni
      summary: Otestovat všechna připojení organizace
      description: Implementováno v aplikaci (přihlášení, X-Cron-Secret nebo podpis HMAC z n8n). Volá ji hodinový hlídač connection-health.json v n8n, proto přijímá i podpis HMAC. Smyslem je zachytit vypršelý token dřív, než se o něj opře publikace.
      security:
        - cookieSession: []
        - webhookHmac: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                organizationId:
                  type: string
                  format: uuid
                  description: Prázdné u podepsaného volání znamená všechny organizace.
                duvod:
                  type: string
      responses:
        "200":
          description: Výsledky testů.
          content:
            application/json:
              schema:
                type: object
                properties:
                  otestovano:
                    type: integer
                  problemy:
                    type: array
                    description: Klíče připojení, která hlásí potíž.
                    items:
                      type: string
                  vysledky:
                    type: array
                    items:
                      type: object
                      properties:
                        connectionId:
                          type: string
                          format: uuid
                        providerKey:
                          type: string
                        ok:
                          type: boolean
                        message:
                          type: string
        "401":
          $ref: "#/components/responses/Nepovoleno"
...
  /api/v1/analytika/synchronizovat:
    post:
      tags:
        - Analytika
      operationId: synchronizovatMetriky
      summary: Stáhnout metriky od poskytovatele
      description: Implementováno v aplikaci (přihlášení, X-Cron-Secret nebo podpis HMAC z n8n). Volá ji denní workflow metrics-sync.json v n8n, proto přijímá i podpis HMAC. Projde publikace, u kterých má smysl se ptát, a uloží nový snímek metrik.
      security:
        - cookieSession: []
        - webhookHmac: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                organizationId:
                  type: string
                  format: uuid
                venueId:
                  type: string
                  format: uuid
                den:
                  type: string
                  format: date
                duvod:
                  type: string
      responses:
        "200":
          description: Hotovo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  aktualizovano:
                    type: integer
                    description: Počet publikací, u kterých přibyl nový snímek.
        "401":
          $ref: "#/components/responses/Nepovoleno"
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: ftm_demo
      description: Přihlášený uživatel. V demo režimu podepsaná cookie ftm_demo, v ostrém provozu cookie Supabase Auth sb-<projekt>-auth-token. Na přístup se vždycky ptá jedno místo, funkce marketing.has_access, a druhou obrannou linií je Row Level Security v databázi.
    cronSecret:
      type: apiKey
      in: header
      name: X-Cron-Secret
      description: Servisní volání z cronu nebo z n8n. Hodnota je CRON_SECRET aplikace, a když není nastavený, APP_SECRET. Porovnává se v konstantním čase.
    webhookHmac:
      type: apiKey
      in: header
      name: X-FoodTab-Signature
      description: Podpis hex(HMAC-SHA256(secret, timestamp + "." + telo)) doplněný hlavičkami X-FoodTab-Timestamp a X-FoodTab-Idempotency-Key. Značka starší nebo novější než 5 minut se odmítne. Klíč se nikdy neukládá v čitelné podobě.
  parameters:
    IdVCeste:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Nepovoleno:
      description: Chybí přihlášení nebo servisní tajemství.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Chyba"
    Zakazano:
      description: Na tuhle akci nemáte oprávnění.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Chyba"
    Nenalezeno:
      description: Záznam neexistuje, nebo do něj nevidíte.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Chyba"
...
  schemas:
    Chyba:
      type: object
      description: Jediný tvar chyby v celém API. Hláška je česká a nikdy neobsahuje tajemství ani vnitřní cesty.
      required:
        - error
      properties:
        error:
          type: string
          examples:
            - Na tuhle akci nemáte oprávnění.
...
    AiNavrh:
      type: object
      description: Ověřený návrh od modelu. Tvar hlídá schéma v lib/providers/ai/schema.ts, nevalidní odpověď se zahodí.
      properties:
        varianty:
          type: array
          items:
            type: object
            additionalProperties: true
        vysvetleni:
          type: string
      additionalProperties: true
...